# firmware: check and send position updates every 5ms, not 10ms