Before:
  Z step = 0.001 --> 41 ms per move
  Z step = -0.001 --> 127 ms per move

After:
  Z step = 0.001 --> 37 ms per move
  Z step = -0.001 --> 112 ms per move

This one is a freebie, but I haven't looked into whether or not doubling the frequency of position checks and updates can cause actual issues.  Our packets are so small that I sorta doubt it, but I haven't actually investigated.  So there's some risk to just going for it with this one.

Tested by: using the `tools/stage_timing.py` script.